+++ packages/serde-toml/src/decode.ml
@@ -84,7 +84,7 @@ and array_backend: 'value. state -> 'value De.t -> 'value array = fun state deco
     values;
   Array.from_list (List.rev !items)
 
-and map_backend: 'value. state -> 'value De.t -> (string * 'value) vec = fun state decode ->
+and dict_backend: 'value. state -> 'value De.t -> (string * 'value) vec = fun state decode ->
   let values = expect_table state.current in
   let result = Vector.create () in
   Document.iter_table
@@ -198,7 +198,7 @@ and backend: state De.backend = {
   option = option_backend;
   list = list_backend;
   array = array_backend;
-  map = map_backend;
+  dict = dict_backend;
   record = record_backend;
   record_mut = record_mut_backend;
   variant = variant_backend;